Sprint2/requirements.txt
Django==2.1.5
psycopg2==2.7.7
psycopg2-binary==2.7.7
django-filter
